transport-service/requirements.txt
PyYAML==5.1.2
PyMySQL==0.9.3
gunicorn==20.0.4
Flask==1.1.1
